Database Reactions.
I have created the database function but they are not complete.
I have created another file array in store in order to send database.
We were only saving the file names and sizes before.
We would will need something more for database.
Now we have an array in state that includes our files.
We can later try to send these files to the database.