A generic Filter.
I have created a new filter for page inputDocuments.vue.
But it can be used in any page. I have created a new file and folder to store
filters. I have connected them in the main.js. So everyone can use it and create their own.

My filter is simple. It converts the given byte value to MB or GB.
There is an explanation about how to use it.
It is in store/tools/Filters.js