Load Input/Output documents mainly fixed + Reflexion classified by 3 tabs  in Backend
Problem is with the phaseID which is unknown, because it is not created in drupal backend "Phase Vorgehensmodell", which contains also the projectID (Projectname). Because it is missing in the backend, the wrong phaseid is displayed when creating a reflexion and in the console, nothing is displayed about the phaseid or projectid, when loading from the startpage to the reflexionpage by choosing a project and the reflexionphase, that don't exist in drupal backend.